requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.23.1
sqlmodel==0.0.8
python-dotenv==1.0.0
mysql-connector-python==8.1.0
SQLAlchemy==1.4.41

#seguridad
passlib[bcrypt]
python-jose==3.3.0
